requirements.txt
aiogram==3.4.1         # июль-2025
python-dotenv>=1.0.0